refactor(editor): little refactor + fix typo in pressDelete tests
