+++ package.json
@@ -36,7 +36,6 @@
     "@types/jasmine": "^2.8.8",
     "@types/webdriverio": "^4.13.0",
     "angular-mocks": "^1.3.20",
-    "http-server": "^0.11.1",
     "jasmine": "^2.8.0",
     "karma": "^3.0.0",
     "karma-browserstack-launcher": "^1.3.0",
@@ -54,7 +53,6 @@
   "scripts": {
     "test": "karma start",
     "test:compatibility": "karma start karma-compatibility.conf.js",
-    "test:serve": "npx http-server test/public",
     "test:selenium": "npx wdio"
   },
   "repository": {
