+++ src/ts/editor/onPressDelete.ts
