+++ test/po/editor.po.ts
@@ -1,11 +1,6 @@
 export default new class EditorPo {
     private selector = `.drawing-zone`;
 
-    navigateToPage(): void {
-        browser.url('/editor');
-        this.editor.waitForVisible();
-    }
-
     waitForVisible(): WebdriverIO.Client<boolean> {
         return browser.waitForVisible(this.selector);
     }
@@ -13,6 +8,4 @@ export default new class EditorPo {
     get editor(): WebdriverIO.Client<WebdriverIO.Element> {
         return browser.element<WebdriverIO.Element>(this.selector);
     }
-
-
 }
