--- test/public/editor.html
@@ -1,6 +1,0 @@
-<html>
-<head></head>
-<body>
-<div style="width: 300px; height: 300px;" contenteditable="true"></div>
-</body>
-</html>
